refactors call, commit_entry and debug to own files
